python/hub-clients/requirements.txt
Twisted==19.7.0
fire==0.2.1
python-memcached==1.59